Comments
Added comments and made some sections more readable.
Removed the mode visibilities from Fit_Info as they weren't used anywhere and were a duplicate of what is in the Run_Settings class.
Removed the Run_Info class (which is now in dedicated file run_settings.py and is renamed the more-appropriate Run_Settings.